reverts stdweb-derive, moves js macro to stdweb-internal-macros
