+++ stdweb-derive/Cargo.toml
@@ -23,4 +23,4 @@ proc-macro2 = "0.4"
 [dependencies.syn]
 version = "0.15"
 default-features = false
-features = ["derive", "parsing", "printing", "extra-traits"]
+features = ["derive", "parsing", "printing"]
