+++ stdweb-internal-macros/Cargo.toml
@@ -27,7 +27,7 @@ proc-macro2 = "0.4"
 [dependencies.syn]
 version = "0.15"
 default-features = false
-features = ["full", "parsing", "printing", "clone-impls"]
+features = ["full", "extra-traits"]
 
 [build-dependencies]
 rustc_version = "0.2"
