+++ stdweb-internal-macros/src/macro_js.rs
@@ -8,12 +8,12 @@ use quote::ToTokens;
 use proc_macro2::TokenTree;
 use proc_macro2::Delimiter;
 
-fn extract_js(cursor: Cursor) -> Result<(Js, Cursor)> {
+fn extract_js<'a>(cursor: Cursor<'a>, block_counter: &mut usize) -> Result<(Js, Cursor<'a>)> {
   let mut script = String::new();
   let mut blocks = Vec::new();
 
   let mut rest = cursor;
-  let mut block_counter = 0;
+//  let mut block_counter = 1;
   let mut next_is_block = false;
 
   while let Some((tt, next)) = rest.token_tree() {
@@ -24,10 +24,11 @@ fn extract_js(cursor: Cursor) -> Result<(Js, Cursor)> {
         Err(err) => return Err(err),
       });
 
-      script += "$";
+      script += "($";
       script += &block_counter.to_string();
+      script += ")";
 
-      block_counter += 1;
+      *block_counter += 1;
       next_is_block = false;
 
       rest = next;
@@ -50,7 +51,7 @@ fn extract_js(cursor: Cursor) -> Result<(Js, Cursor)> {
         let (group, _, next) = rest.group(group.delimiter()).unwrap();
         rest = next;
 
-        let mut js = extract_js(group)?.0;
+        let mut js = extract_js(group, block_counter)?.0;
 
         script += start;
         script += &js.script;
@@ -60,8 +61,14 @@ fn extract_js(cursor: Cursor) -> Result<(Js, Cursor)> {
         continue;
       },
       _ => {
-        script += &tt.to_string();
-        script += " ";
+        let token = tt.to_string();
+        let token_alpha =  token.chars().next().unwrap().is_alphanumeric();
+
+        if token_alpha && script.chars().last().unwrap_or(' ').is_alphanumeric() {
+          script += " ";
+        }
+
+        script += &token;
       }
     }
 
@@ -83,7 +90,7 @@ pub struct Js {
 impl Parse for Js {
   fn parse(input: ParseStream) -> Result<Self> {
     input.step(|cursor| {
-      extract_js(*cursor)
+      extract_js(*cursor, &mut 1)
     })
   }
 }
@@ -94,4 +101,43 @@ pub fn transform_js(input: Js) -> TokenStream {
   let script_str = script.to_string();
 
   quote!(__js_raw_asm!(#script_str, #(#blocks)*))
+}
+
+#[cfg(test)]
+mod tests {
+  use syn::parse2;
+  use proc_macro2::TokenStream;
+  use macro_js::Js;
+  use std::str::FromStr;
+
+  fn parse(tokens: &str) -> String {
+    let ts = TokenStream::from_str(tokens).unwrap();
+    let js: Js = parse2(ts).unwrap();
+    js.script
+  }
+
+  #[test]
+  fn stringify() {
+    assert_eq!(parse("return thing;"), "return thing;");
+    assert_eq!(parse("console.log"), "console.log");
+    assert_eq!(parse("1.0"), "1.0");
+    assert_eq!(parse("[ 1.0 ]"), "[1.0]");
+    assert_eq!(parse("{ 1.0 }"), "{1.0}");
+    assert_eq!(parse("( 1.0 )"), "(1.0)");
+    assert_eq!(parse("a b"), "a b");
+    assert_eq!(parse("==="), "===");
+    assert_eq!(parse("++i"), "++i");
+    assert_eq!(parse("i++"), "i++");
+    assert_eq!(parse("--i"), "--i");
+    assert_eq!(parse("i--"), "i--");
+    assert_eq!(parse("( @{1} ); ").replace(" ", ""), "(($1));");
+    assert_eq!(
+      parse("console.log(\"Hello!\", @{1234i32} );").replace(" ", ""),
+      "console.log(\"Hello!\",($1));"
+    );
+    assert_eq!(
+      parse("@{a}.fn( @{b} );").replace(" ", ""),
+      "($1).fn(($2));"
+    );
+  }
 }
